Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 int main(){
-    int a;
-    scanf("%d",&a);
-    printf("Area: %d",3.14(a**2));
+    float a;
+    scanf("%f",&a);
+    printf("Area: %f",3.14(a**2));
     return 0;
 }
